:sparkles: Add loading and watching as core part of the repository API
No need to externally pass in the loaders to the repository.
